+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Bug Fix
 - Fixed a bug that prevented project viewers from searching for subjects and viewing subject events
+- Fixed an issue saving email preferences on user settings page
 
 ## 0.31.1 (November 12, 2015)